Feat: add error boundaries (#30)
<!-- Please provide enough information so that others can review your
pull request. -->
<!-- Keep pull requests small and focused on a single change. -->

### Summary

Added basic ErrorBoundary component in mini-apps and host & shell app.
Used it in places near React.Suspense.

In some cases the navigation bar is present so there's a chance to go
back to the previous screen, but in most cases where we import the whole
mini-app the NavBar is hidden, and we end up with no way to recover from
the ErrorBoundary fallback screen. To keep the experience smooth we
should consider adding a 'Go Back' button to go to the previous screen
in case of an error.

There is also some repetition involved with ErrorBoundary being copied
over to the mini-apps, however this conforms to the pattern present in
the repository. In theory implementation could be different in each of
the mini-apps but isn't for the sake of simplicity.

<!-- What existing problem does the pull request solve? Can you solve
the issue with a different approach? -->